Data Analysis Example with a Quick Random Forest Classifier
Data Analysis.pdf is the main Analysis document, showing the data in detail.
The Jupyter Notebook Analysis has the in depth analysis of the Data - essentially a duplicate of Analysis.pdf, with the difference of including the chosen variables used in the Prediction Modelling Notebook.
The Prediction Modelling Notebook contains a simple Random Forest Classifier.